addDays method removed for simplicity
The use of addDays was the original source of contention with the bug
residing in sending days and comparing with self.days, creating pointers
to the same bit of memory and creating compatibility issues. The fix for
this ended up negating the need for the method entirely, so it was
removed and replaced with two lines of code that simply used the
new collection of days, created when a UI change was triggered.