Removed an array count and added an maxCycleCount
Since the API only creates information as needed, a received array
from a new profile will return 2 items:

- the day of the first cycle input during profile creation
- the current day.

This count was used both when deciding how many days to display
and where the day line should be drawn. A conditional solved the
first issue and the use of cycleDay instead of days.count resolved
the second.

A static variable called maxCycleCount is implemented so that a cycle
range of over 30 will maintain when moving backwards. Conditionals
were put in place to adjust to a cycle change and an array count of
day entities was no longer needed.